More debugging.
Trying out MSCOCO2014 with the weights from PJReddie.
This is not generating meaningful inferences at all.
The box locations and class labels make no sense.

Next: Try running DarkNet, or use a different implementation
of YOLO3.

Could be that the code I'm using for YOLO3 is just buggy.
Or the weights file is not correct for the model architecture.

Alternately use Faster-RCNN or some other model architecture
which has a better PyTorch implementation.